+++ packages/fireworks/src/fireworks-provider.ts
@@ -8,12 +8,13 @@ import {
   EmbeddingModelV3,
   ImageModelV3,
   LanguageModelV2,
-  ProviderV2,
+  ProviderV3,
 } from '@ai-sdk/provider';
 import {
   FetchFunction,
   loadApiKey,
   withoutTrailingSlash,
+  withUserAgentSuffix,
 } from '@ai-sdk/provider-utils';
 import { z } from 'zod/v4';
 import { FireworksChatModelId } from './fireworks-chat-options';
@@ -21,6 +22,7 @@ import { FireworksCompletionModelId } from './fireworks-completion-options';
 import { FireworksEmbeddingModelId } from './fireworks-embedding-options';
 import { FireworksImageModel } from './fireworks-image-model';
 import { FireworksImageModelId } from './fireworks-image-options';
+import { VERSION } from './version';
 
 export type FireworksErrorData = z.infer<typeof fireworksErrorSchema>;
 
@@ -54,7 +56,7 @@ or to provide a custom fetch implementation for e.g. testing.
   fetch?: FetchFunction;
 }
 
-export interface FireworksProvider extends ProviderV2 {
+export interface FireworksProvider extends ProviderV3 {
   /**
 Creates a model for text generation.
 */
@@ -99,14 +101,18 @@ export function createFireworks(
   options: FireworksProviderSettings = {},
 ): FireworksProvider {
   const baseURL = withoutTrailingSlash(options.baseURL ?? defaultBaseURL);
-  const getHeaders = () => ({
-    Authorization: `Bearer ${loadApiKey({
-      apiKey: options.apiKey,
-      environmentVariableName: 'FIREWORKS_API_KEY',
-      description: 'Fireworks API key',
-    })}`,
-    ...options.headers,
-  });
+  const getHeaders = () =>
+    withUserAgentSuffix(
+      {
+        Authorization: `Bearer ${loadApiKey({
+          apiKey: options.apiKey,
+          environmentVariableName: 'FIREWORKS_API_KEY',
+          description: 'Fireworks API key',
+        })}`,
+        ...options.headers,
+      },
+      `ai-sdk/fireworks/${VERSION}`,
+    );
 
   interface CommonModelConfig {
     provider: string;