+++ packages/openai-compatible/src/openai-compatible-provider.ts
@@ -2,7 +2,7 @@ import {
   EmbeddingModelV3,
   ImageModelV3,
   LanguageModelV2,
-  ProviderV2,
+  ProviderV3,
 } from '@ai-sdk/provider';
 import {
   FetchFunction,
@@ -24,7 +24,7 @@ export interface OpenAICompatibleProvider<
   COMPLETION_MODEL_IDS extends string = string,
   EMBEDDING_MODEL_IDS extends string = string,
   IMAGE_MODEL_IDS extends string = string,
-> extends Omit<ProviderV2, 'imageModel'> {
+> extends Omit<ProviderV3, 'imageModel'> {
   (modelId: CHAT_MODEL_IDS): LanguageModelV2;
 
   languageModel(