+++ packages/vercel/src/vercel-provider.ts
@@ -1,15 +1,17 @@
 import {
   LanguageModelV2,
   NoSuchModelError,
-  ProviderV2,
+  ProviderV3,
 } from '@ai-sdk/provider';
 import { OpenAICompatibleChatLanguageModel } from '@ai-sdk/openai-compatible';
 import {
   FetchFunction,
   loadApiKey,
   withoutTrailingSlash,
+  withUserAgentSuffix,
 } from '@ai-sdk/provider-utils';
 import { VercelChatModelId } from './vercel-chat-options';
+import { VERSION } from './version';
 
 export interface VercelProviderSettings {
   /**
@@ -31,7 +33,7 @@ or to provide a custom fetch implementation for e.g. testing.
   fetch?: FetchFunction;
 }
 
-export interface VercelProvider extends ProviderV2 {
+export interface VercelProvider extends ProviderV3 {
   /**
 Creates a model for text generation.
 */
@@ -49,14 +51,18 @@ export function createVercel(
   const baseURL = withoutTrailingSlash(
     options.baseURL ?? 'https://api.v0.dev/v1',
   );
-  const getHeaders = () => ({
-    Authorization: `Bearer ${loadApiKey({
-      apiKey: options.apiKey,
-      environmentVariableName: 'VERCEL_API_KEY',
-      description: 'Vercel',
-    })}`,
-    ...options.headers,
-  });
+  const getHeaders = () =>
+    withUserAgentSuffix(
+      {
+        Authorization: `Bearer ${loadApiKey({
+          apiKey: options.apiKey,
+          environmentVariableName: 'VERCEL_API_KEY',
+          description: 'Vercel',
+        })}`,
+        ...options.headers,
+      },
+      `ai-sdk/vercel/${VERSION}`,
+    );
 
   interface CommonModelConfig {
     provider: string;