+++ packages/azure/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @ai-sdk/azure
 
+## 2.1.0-beta.3
+
+### Patch Changes
+
+- Updated dependencies [2e86082]
+  - @ai-sdk/openai@2.1.0-beta.3
+
+## 2.1.0-beta.2
+
+### Patch Changes
+
+- 0c4822d: feat: `EmbeddingModelV3`
+- 1cad0ab: feat: add provider version to user-agent header
+- Updated dependencies [4920119]
+- Updated dependencies [0c4822d]
+- Updated dependencies [1cad0ab]
+  - @ai-sdk/openai@2.1.0-beta.2
+  - @ai-sdk/provider@2.1.0-beta.1
+  - @ai-sdk/provider-utils@3.1.0-beta.2
+
 ## 2.1.0-beta.1
 
 ### Patch Changes